Colour for CSPA expiry date working
The colouring for the CSPA membership expiry date is lower than today’s
date is working.

Unfortunately unable to get the controller to keep to limit the packers
that are on the team AND CSPA membership exert date lower than today’s
date.

@packers = Packer.find(:all, :conditions => { :p_team => "t" ,
:p_cspa_date => 2.years.ago }, :order => "p_name")

Works nicely….unable to change to LESS THAN.